Move GFileUtils to base-services
