+++ subprojects/base-services/src/main/java/org/gradle/util/internal/LimitedDescription.java
