+++ subprojects/base-services/src/test/groovy/org/gradle/util/GUtilTest.groovy
