+++ subprojects/core/src/main/java/org/gradle/api/internal/project/taskfactory/OutputDirectoryPropertyAnnotationHandler.java
@@ -18,7 +18,7 @@ package org.gradle.api.internal.project.taskfactory;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.tasks.OutputDirectory;
 import org.gradle.api.tasks.TaskOutputFilePropertyBuilder;
-import org.gradle.util.GFileUtils;
+import org.gradle.util.DeferredUtil;
 
 import java.io.File;
 import java.lang.annotation.Annotation;
@@ -57,7 +57,7 @@ public class OutputDirectoryPropertyAnnotationHandler extends AbstractOutputProp
     }
 
     private File toFile(Object value) {
-        Object unpacked = GFileUtils.unpack(value);
+        Object unpacked = DeferredUtil.unpack(value);
         if (unpacked instanceof Path) {
             return ((Path) unpacked).toFile();
         }
