+++ subprojects/core/src/main/java/org/gradle/api/internal/project/taskfactory/OutputFilePropertyAnnotationHandler.java
@@ -18,7 +18,7 @@ package org.gradle.api.internal.project.taskfactory;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.tasks.OutputFile;
 import org.gradle.api.tasks.TaskOutputFilePropertyBuilder;
-import org.gradle.util.GFileUtils;
+import org.gradle.util.DeferredUtil;
 
 import java.io.File;
 import java.lang.annotation.Annotation;
@@ -55,7 +55,7 @@ public class OutputFilePropertyAnnotationHandler extends AbstractOutputPropertyA
     }
 
     private File toFile(Object value) {
-        Object unpacked = GFileUtils.unpack(value);
+        Object unpacked = DeferredUtil.unpack(value);
         if (unpacked instanceof Path) {
             return ((Path) unpacked).toFile();
         }
