+++ subprojects/core/src/main/java/org/gradle/util/RelativePathUtil.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.util;
+
+import org.gradle.internal.UncheckedException;
+
+import java.io.File;
+
+public class RelativePathUtil {
+    /**
+     * Returns a relative path from 'from' to 'to'
+     *
+     * @param from where to calculate from
+     * @param to where to calculate to
+     * @return The relative path
+     */
+    public static String relativePath(File from, File to) {
+        try {
+            return org.apache.tools.ant.util.FileUtils.getRelativePath(from, to);
+        } catch (Exception e) {
+            throw UncheckedException.throwAsUncheckedException(e);
+        }
+    }
+}
