+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/test/fixtures/server/sftp/SftpResource.groovy
@@ -19,7 +19,7 @@ package org.gradle.test.fixtures.server.sftp
 import org.apache.commons.io.FilenameUtils
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.test.fixtures.resource.RemoteResource
-import org.gradle.util.GFileUtils
+import org.gradle.util.RelativePathUtil
 
 class SftpResource implements RemoteResource {
 
@@ -36,7 +36,7 @@ class SftpResource implements RemoteResource {
     }
 
     String getPathOnServer() {
-        return "/${GFileUtils.relativePath(server.baseDir, file)}"
+        return "/${RelativePathUtil.relativePath(server.baseDir, file)}"
     }
 
     URI getUri() {
